Move JobDefinition-specific functions to job_definition
